All container components with mui
