+++ src/components/common/dialog/content.js
@@ -6,9 +6,11 @@ import classnames from 'classnames'
 import { withStyles } from '@material-ui/core/styles'
 
 const propRow = ({ classes, left, right, className, classNameLeft, classNameRight, style = {} }) =>
-    <Grid item spacing={16} container xs={12} className={classnames(className)} style={style}>
-        <Grid item xs={12} sm={4} lg={3} className={classnames(classes.leftCol, classes.uppercase, classNameLeft)}>{left}</Grid >
-        <Grid item xs={12} sm={8} lg={9} className={classnames(classes.rightCol, classes.breakLong, classNameRight)}>{right}</Grid >
+    <Grid item xs={12} className={classnames(className)} style={style}>
+        <Grid container spacing={16}>
+            <Grid item xs={12} sm={4} lg={3} className={classnames(classes.leftCol, classes.uppercase, classNameLeft)}>{left}</Grid >
+            <Grid item xs={12} sm={8} lg={9} className={classnames(classes.rightCol, classes.breakLong, classNameRight)}>{right}</Grid >
+        </Grid>
     </Grid>
 
 const contentBox = ({ classes, children, className }) =>
