--- src/components/common/floating_btn_progress/theme.css
@@ -1,42 +1,0 @@
-@import './../../../../node_modules/react-toolbox/lib/progress_bar/theme.css';
-
-:root {
-  --progress-height: calc(0.4 * var(--unit));
-  --progress-main-color: var(--color-primary);
-  --progress-secondary-color: color(var(--color-primary-contrast) a(70%));
-  --progress-disabled-color: color(var(--color-black) a(26%));
-  --circle-wrapper-width: 64;
-  --circle-radius: 28;
-  --scale-ratio: calc(var(--circle-radius) / 20);
-}
-
-.button{
-  position: relative;
-}
-
-.circular {
-  position: absolute;
-  top: -4px;
-  left: -4px;
-  right: -4px;
-  bottom: -4px;
-  z-index: 200;
-
-  &.indeterminate {
-    & .circle {
-      animation: circular-indeterminate-bar-rotate 2s linear infinite;
-    }
-
-    & .path {
-      animation: circular-indeterminate-bar-dash 1.5s ease-in-out infinite;
-      stroke-dasharray: calc(var(--scale-ratio) * 1), calc(var(--scale-ratio) * 200);
-      stroke-dashoffset: 0;
-    }
-
-    &.multicolor .path {
-      animation:
-        circular-indeterminate-bar-dash 1.5s ease-in-out infinite,
-        colors calc(1.5s * 4) ease-in-out infinite;
-    }
-  }
-}
