+++ src/components/dashboard/containers/AddItem/styles.js
@@ -1,0 +1,14 @@
+export const styles = theme => {
+    const spacing = theme.spacing.unit
+    return {
+        appBar: {
+            marginBottom: spacing
+        },
+        tabsContainer: {
+            display: 'flex',
+            flexGrow: 1,
+            overflowY: 'auto',
+            position: 'relative'
+        }
+    }
+}
