+++ src/components/dashboard/containers/ImgDialog/styles.js
@@ -1,0 +1,5 @@
+export const styles = theme => {
+    // const spacing = theme.spacing.unit * 3
+    return {
+    }
+}
