+++ src/components/dashboard/containers/ItemIpfsDetails/index.js
@@ -1,0 +1,1 @@
+export { default } from './ItemIpfsDetails'
