--- src/components/dashboard/containers/datepicker.css
@@ -1,50 +1,0 @@
-.campaignPeriodContainer {
-    color: #fff;
-    position: fixed;
-    top: 104px;
-    transform: translateY(-50%);
-    right: 124px;
-    z-index: 110;
-}
-
-.campaignPeriodContainer, .statsDatePicker {
-    & .input, & .inputElement {
-        padding: 0;
-        margin: 0;
-        border: 0;
-        display: inline-block;
-        color: #fff;
-        vertical-align: text-bottom;
-    }
-    & .label {
-        display: none;
-    }
-    & > div {
-        display: inline-block;
-    }
-    & > span {
-        margin-left: 5px;
-        margin-right: 5px;
-        opacity: 0.6;
-        vertical-align: text-bottom;
-    }
-}
-
-.statsDatePicker {
-    display: flex;
-    align-items: center;
-
-    &.active {
-        & .input, & .inputElement, & span {
-            opacity: 1;
-        }
-    }
-    &.button {
-        opacity: 1;
-    }
-
-    & .input, & .inputElement, & span {
-        color: #fff;
-        opacity: 0.6;
-    }
-}
