--- src/components/dashboard/containers/itemDialogTheme.css
@@ -1,10 +1,0 @@
-.tab {
-    margin-top: 20px;
-    position: relative;
-    flex-grow: 1;
-}
-
-.tabs {
-    display: flex;
-    flex-grow: 1;
-}
